requirements.txt
fastapi
python-dotenv
pydantic
pydantic-settings
tavily-python
trafilatura
sentence_transformers
numpy
ollama
fastapi[standard]
langchain
langchain-core
langchain-ollama
langchain-community

# >> # Create a new one
# >> python -m venv .venv
# >> 
# >> # Activate it
# >> .\.venv\Scripts\Activate.ps1
# >> 
# >> # Install packages
# >> pip install -r requirements.txt